Feat: 심화 과제 구현
